requirements.txt
openai
torch
numpy
pandas 
tensorflow
keras
python-dotenv
transformers
pyyaml
pytest
geopy
Nominatim
re
ast
tenacity
jinja2
